Refactor amici.ode_export, add AmiciCxxCodePrinter
Move custom codeprinting to a separate code printer class and reuse the same instance for all printing.
Avoids plenty of reinstantiations of the code printer and keeps all settings in one place.
