+++ python/sdist/amici/cxxcodeprinter.py
@@ -1,0 +1,1 @@
+../../amici/cxxcodeprinter.py
